requirements.txt
numpy
scipy
scikit-learn
h5py
pandas
pyyaml
pingouin
notebook
jupyterlab
matplotlib
tqdm

torch>=1.13.1
torchvision>=0.14.1
umap-learn==0.5
tensorflow-macos
tensorflow-probability
